Add place for checkResults in client data store
