+++ packages/apollo-mst/src/CheckResult.ts
@@ -1,0 +1,12 @@
+import { types } from 'mobx-state-tree'
+
+import { AnnotationFeature } from './AnnotationFeature'
+
+export const CheckResult = types.model('CheckResult', {
+  id: types.identifier,
+  name: types.string,
+  ids: types.array(types.safeReference(AnnotationFeature)),
+  location: types.model({ start: types.number, end: types.number }),
+  ignored: false,
+  message: types.string,
+})
